+++ appconf/tests/settings.py
@@ -1,0 +1,3 @@
+SIMPLE_VALUE = True
+
+CONFIGURED_VALUE = 'wrong'